+++ modules/Grades/ReportCardComments.php
@@ -180,7 +180,10 @@ if ( ! $_REQUEST['modfunc'] )
 
 		if ( empty( $subjects_RET ) )
 		{
-			$subject_select .= '<option value="">' . sprintf( _( 'No %s were found.' ), ngettext( 'Course', 'Courses', 0 ) ) . '</option>';
+			$subject_select .= '<option value="">' . sprintf(
+				_( 'No %s were found.' ),
+				mb_strtolower( ngettext( 'Course', 'Courses', 0 ) )
+			) . '</option>';
 		}
 		else
 		{
@@ -204,7 +207,10 @@ if ( ! $_REQUEST['modfunc'] )
 
 		if ( empty( $courses_RET ) )
 		{
-			$course_select .= '<option value="">' . sprintf( _( 'No %s were found.' ), ngettext( 'Course', 'Courses', 0 ) ) . '</option>';
+			$course_select .= '<option value="">' . sprintf(
+				_( 'No %s were found.' ),
+				mb_strtolower( ngettext( 'Course', 'Courses', 0 ) )
+			) . '</option>';
 		}
 		else
 		{